refactor: Migrate AI workflows and notifications to DBOS
