+++ apps/web/src/server.ts
@@ -1,32 +1,10 @@
 import "@/integrations/otel/init";
-import "@/features/ai/backfill-keywords.workflow";
+import "@/integrations/dbos/workflows/backfill-keywords.workflow";
 
-import { DBOS } from "@dbos-inc/dbos-sdk";
 import handler, { createServerEntry } from "@tanstack/react-start/server-entry";
-import { env } from "@core/environment/server";
-import { getLogger } from "@core/logging/root";
+import { launchDBOS } from "@/integrations/dbos/init";
 
-const logger = getLogger().child({ module: "web-server" });
-
-DBOS.setConfig({
-   name: "montte-web",
-   systemDatabaseUrl: env.DATABASE_URL,
-   logLevel: env.LOG_LEVEL ?? "info",
-});
-
-DBOS.launch()
-   .then(() => {
-      logger.info("DBOS runtime started");
-   })
-   .catch((err) => {
-      logger.error({ err }, "DBOS launch failed");
-   });
-
-const shutdown = async () => {
-   await DBOS.shutdown();
-};
-process.on("SIGTERM", () => void shutdown());
-process.on("SIGINT", () => void shutdown());
+launchDBOS();
 
 export default createServerEntry({
    fetch(request) {
