+++ packages/notifications/src/schema.ts
@@ -4,8 +4,8 @@ export const jobNotificationSchema = z.object({
    jobId: z.string(),
    type: z.string(),
    status: z.union([z.literal("completed"), z.literal("failed")]),
+   message: z.string(),
    payload: z.record(z.string(), z.unknown()).optional(),
-   error: z.string().optional(),
    teamId: z.string().uuid(),
    timestamp: z.string().datetime(),
 });
